Move useGetBranding hook to be in a less specific subfolder
