+++ app/src/app/components/contents/common/hooks/useGetBranding.ts
@@ -1,7 +1,7 @@
 import useSWR from "swr";
-import appConfig from "../../../../config/appConfig";
-import { fetcher } from "../../../../lib/fetch";
-import { BrandingConfiguration } from "../../../../types";
+import appConfig from "../../../../../config/appConfig";
+import { fetcher } from "../../../../../lib/fetch";
+import { BrandingConfiguration } from "../../../../../types";
 
 export const useGetBranding = () => {
   const { data, isLoading, error } = useSWR<BrandingConfiguration>(
